Update 5_diagnostics_plot.R Line 110 Names of prop Variable
prop sum of loads had unclean names. In line 110 you wanted to change these names using the clean_names variable. The indexing in line 110 seems to be taking the length of convars to length of cleannames, this only displays 1 name. I'm not sure what the purpose of this index was. Eliminating the indexing and using just the clean_names variable seems to work.